# docs/openapi.yaml
...
openapi: 3.0.1
info:
  title: SEGS API
  description: 'Secure Email Gateway Suite HTTP API. Authenticate with a stateful JWT (``Authorization:
    Bearer``) or the ``seg_session`` cookie. SCIM 2.0 is at ``/scim/v2``. Interactive ``/docs``
    is not served; the contract is ``docs/openapi.yaml`` and is what API Gateway imports.'
  version: 1.0.0
  x-generated-by: backend.api.openapi
servers:
- url: /
  description: Same origin as the SOC console (CloudFront)
paths:
...
  /api/feed:
    get:
      summary: Get Feed
      operationId: get_feed_api_feed_get
      parameters:
      - name: verdict
        in: query
        required: false
        schema:
          type: string
          description: 'Overview tile: safe, suspicious, or malicious.'
          default: ''
          title: Verdict
        description: 'Overview tile: safe, suspicious, or malicious.'
      - name: origin
        in: query
        required: false
        schema:
          type: string
          description: ISO country from the Overview origin map.
          default: ''
          title: Origin
        description: ISO country from the Overview origin map.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
      - cookieAuth: []
  /api/feed/search:
    post:
      summary: Spotlight Search
      description: 'Natural-language spotlight search: LLM → validated SQL → mail rows.'
      operationId: spotlight_search_api_feed_search_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpotlightSearchBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
      - cookieAuth: []
  /api/feed/item/{queue_id}:
    get:
      summary: Get Feed Item
      description: One copy (and its thread) even when it has aged off the 500-row feed page.
      operationId: get_feed_item_api_feed_item__queue_id__get
      parameters:
      - name: queue_id
        in: path
        required: true
        schema:
          type: string
          title: Queue Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
      - cookieAuth: []
  /api/feed/refresh:
    post:
      summary: Refresh Feed
      operationId: refresh_feed_api_feed_refresh_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - bearerAuth: []
      - cookieAuth: []
  /api/feed/retry-ai:
    post:
      summary: Retry Ai
      operationId: retry_ai_api_feed_retry_ai_post
      requestBody:
        content:
          application/json:
            schema:
              title: Body
              $ref: '#/components/schemas/RetryAiBody'
              nullable: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
      - cookieAuth: []
  /api/audit:
    get:
      summary: Get Audit
      description: Gateway shadow decisions + console activity (logins, user admin, triage).
      operationId: get_audit_api_audit_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - bearerAuth: []
      - cookieAuth: []
  /api/audit/me:
    get:
      summary: Get My Audit
      description: The signed-in user's own console activity (no gateway shadow decisions).
      operationId: get_my_audit_api_audit_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - bearerAuth: []
      - cookieAuth: []
  /api/activity/email-view:
    post:
      summary: Record Email View
      description: Analyst opened or left an email details page, or saved the original file.
      operationId: record_email_view_api_activity_email_view_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailViewBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
      - cookieAuth: []
  /api/quarantine/{queue_id}/release:
    post:
      summary: Release
      operationId: release_api_quarantine__queue_id__release_post
      parameters:
      - name: queue_id
        in: path
        required: true
        schema:
          type: string
          title: Queue Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
      - cookieAuth: []
  /api/quarantine/{queue_id}/keep-blocked:
    post:
      summary: Keep Blocked
      operationId: keep_blocked_api_quarantine__queue_id__keep_blocked_post
      parameters:
      - name: queue_id
        in: path
        required: true
        schema:
          type: string
          title: Queue Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
      - cookieAuth: []
  /api/quarantine/{queue_id}/reevaluate:
    post:
      summary: Reevaluate
      operationId: reevaluate_api_quarantine__queue_id__reevaluate_post
      parameters:
      - name: queue_id
        in: path
        required: true
        schema:
          type: string
          title: Queue Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
      - cookieAuth: []
  /api/quarantine/{queue_id}/download:
    get:
      summary: Download
      operationId: download_api_quarantine__queue_id__download_get
      parameters:
      - name: queue_id
        in: path
        required: true
        schema:
          type: string
          title: Queue Id
      - name: intent
        in: query
        required: false
        schema:
          type: string
          description: view = load into the console; omit to log a file download
          default: ''
          title: Intent
        description: view = load into the console; omit to log a file download
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
      - cookieAuth: []
...
components:
  schemas:
...
    EmailViewBody:
      properties:
        queue_id:
          type: string
          maxLength: 200
          minLength: 1
          title: Queue Id
        event:
          type: string
          enum:
          - open
          - leave
          - download
          title: Event
          default: leave
        dwell_ms:
          type: integer
          maximum: 86400000.0
          minimum: 0.0
          title: Dwell Ms
          default: 0
        subject:
          type: string
          maxLength: 500
          title: Subject
          default: ''
        from_addr:
          type: string
          maxLength: 320
          title: From Addr
          default: ''
      type: object
      required:
      - queue_id
      title: EmailViewBody
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RetryAiBody:
      properties:
        queue_ids:
          title: Queue Ids
          items:
            type: string
          type: array
          maxItems: 100
          nullable: true
      type: object
      title: RetryAiBody
...
    SpotlightSearchBody:
      properties:
        q:
          type: string
          maxLength: 500
          minLength: 1
          title: Q
        verdict:
          type: string
          maxLength: 32
          title: Verdict
          default: ''
      type: object
      required:
      - q
      title: SpotlightSearchBody
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Stateful JWT (RFC 6750). Same token as the seg_session cookie; jti is stored
        server-side.
    cookieAuth:
      type: apiKey
      in: cookie
      name: seg_session
      description: Session cookie set at login. Same credential as Bearer.
    scimBearer:
      type: http
      scheme: bearer
      description: SCIM 2.0 bearer (SEG_SCIM_BEARER_TOKEN) or an admin JWT.